This is a complete machine learning end-to-end project for building a Web App for the detection of Spam messages or the classification of any messages into Spam and Ham with very high accuracy. For this project, NLP Tools (NLTK, Gensim etc.), Bag of words model, TFIDF and classification model (SVM) are used. For building the Web App, streamlit library is used. We productionized the enire machine learning pipeline, created Web API with front end UI in streamlit, created docker image of the enire application and at the end deployed the containerized ML application onto Google Kubernetes Engine (GKE) in GCP and exposed it to internet and finally tested this ML App from browsers and predicted spam successfully for a single text message (just typed in the box in App) and also multiple different text messages (read from a csv file).
This is a complete machine learning end-to-end project for building a Web App for the detection of Spam messages or the classification of any messages into Spam and Ham with very high accuracy. For this project, NLP Tools (NLTK, Gensim etc.), Bag of words model, TFIDF and classification model (SVM) are used. For building the Web App, streamlit library is used. We productionized the enire machine learning pipeline, created Web API with front end UI in streamlit, created docker image of the enire application and at the end deployed the containerized ML application onto Google Kubernetes Engine (GKE) in GCP and exposed it to internet and finally tested this ML App from browsers and predicted spam successfully for a single text message (just typed in the box in App) and also multiple different text messages (read from a csv file).